requirements.txt
discord.py
aiohttp
apscheduler